Got basic DB up and running

+++ src/main/kotlin/net/danlew/predictit/api/model/ApiContract.kt
@@ -3,13 +3,14 @@ package net.danlew.predictit.api.model
 import com.squareup.moshi.JsonClass
 import net.danlew.predictit.model.Contract
 import net.danlew.predictit.model.ContractId
+import net.danlew.predictit.model.Price
 import java.math.BigDecimal
 
 @JsonClass(generateAdapter = true)
 internal data class ApiContract(
   val id: Long,
   val name: String,
-  val lastTradePrice: BigDecimal
+  val lastTradePrice: Double
 ) {
 
   fun toContract() = Contract(
